makes app ts and removes all the customisations to get the base to work
