+++ examples/with-localstorage/package.json
@@ -3,6 +3,8 @@
     "version": "1.0.0",
     "private": true,
     "dependencies": {
+        "@types/cors": "^2.8.12",
+        "@types/set-cookie-parser": "^2.4.2",
         "axios": "^0.21.0",
         "cors": "^2.8.5",
         "express": "^4.17.1",
@@ -13,9 +15,10 @@
         "react-dom": "^18.0.0",
         "react-router-dom": "^6.2.1",
         "react-scripts": "4.0.0",
-        "set-cookie-parser": "^2.4.8",
-        "supertokens-auth-react": "^0.22.3",
-        "supertokens-node": "^9.2.0",
+        "set-cookie-parser": "^2.5.0",
+        "supertokens-auth-react": "^0.23.0",
+        "supertokens-node": "^9.3.0",
+        "ts-node-dev": "^2.0.0",
         "web-vitals": "^0.2.4"
     },
     "devDependencies": {
@@ -30,8 +33,8 @@
         "build": "react-scripts build",
         "test": "react-scripts test",
         "eject": "react-scripts eject",
-        "api-server": "node api-server.js",
-        "api-server:dev": "nodemon api-server.js",
+        "api-server": "ts-node-dev --project api-server/tsconfig.json api-server/index.ts",
+        "api-server:dev": "ts-node-dev --project api-server/tsconfig.json api-server/index.ts",
         "server": "node server.js",
         "server:dev": "nodemon server.js",
         "dev": "npm-run-all --parallel spa api-server:dev",
