+++ examples/with-localstorage/src/Footer/index.tsx
