+++ examples/with-localstorage/src/Home/CallAPIView.tsx
