+++ examples/with-localstorage/src/Home/Logout.tsx
@@ -1,4 +1,4 @@
-export default function Logout(props) {
+export default function Logout(props: { logoutClicked: () => void }) {
     let logoutClicked = props.logoutClicked;
 
     return (
