+++ examples/with-localstorage/src/SessionExpiredPopup/index.tsx
